Moved SqlQuery to repo
